Move : http와 관련된 enum 클래스 위치 이동
