+++ src/main/java/webserver/RequestHandler.java
@@ -13,7 +13,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import static dto.HttpResponse.*;
-import static http.Status.*;
+import static http.constants.Status.*;
 
 public class RequestHandler implements Runnable {
     public static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
